Make `hostid mandatory and try to free occupied port
The commit resolves the problem that previously occupied port
cannot be easily freed, a newly connected client with identical
`hostid has to bind to another port number. Also, the commit
requires that `hostid string from clients cannot be empty. If
`hostid is not specified, an random string will be generated.